Add WebSocket listener to the broker.
This patch adds support for using a WebSocket connection with JSON
messages for the broker. The messages are encoded using JSON to make it
easier for webbrowsers to implement the feature for the JavaScript
Snowflake proxy while also making it easy for the go-proxy to be
implemented and share common code with the Broker source code.

See: https://bugs.torproject.org/29736