Add support for authenticating proxies to the broker.
This patch makes the Go Proxy generate a P256 ECDSA public key, which is
used to authenticate to the broker after we have received the nonce
message.

The nonce message is hashed using SHA-2/256 and then signed. The
signature and public key is sent to the broker using the new "proxy
authenticate" message type.

We currently don't use the authentication for anything except storing
the PublicKey as part of the BrokerConnection type, which later will
allow us to use it for different things.

See: https://bugs.torproject.org/29260